services/api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
python-dateutil==2.9.0.post0
pydantic==2.9.2
pandas==2.2.2
openpyxl==3.1.5
apscheduler==3.10.4
cryptography>=42.0.0
